requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
da-vinci==0.4.0
Django==5.0.7
django-allauth==64.0.0
django-ckeditor==6.7.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-mptt==0.16.0
django-resized==1.0.2
django-thumbnails==0.8.0
fixtures==4.1.0
idna==3.7
jwt==1.3.1
pbr==6.0.0
pillow==10.4.0
pycparser==2.22
requests==2.32.3
shortuuid==1.0.13
sorl-thumbnail==12.10.0
sqlparse==0.5.0
thumbnail==1.5
typing_extensions==4.12.2
urllib3==2.2.2